refactor: Replace under package name 'domain' #1
