+++ src/main/java/com/profect/delivery/domain/menu/dto/CreateMenuRequest.java
@@ -1,4 +1,4 @@
-package com.profect.delivery.menu.dto;
+package com.profect.delivery.domain.menu.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
