+++ src/main/java/com/profect/delivery/domain/menu/dto/UpdateMenuResponse.java
@@ -1,4 +1,4 @@
-package com.profect.delivery.menu.dto;
+package com.profect.delivery.domain.menu.dto;
 
 import java.util.UUID;
